update product details
