+++ src/components/ProductDetails/index.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react'
-import { ProductMainDetails } from '../ProductMainDetails'
+import { ProductDetatils as ProductDetatilsController } from 'ordering-components-admin'
+import { ProductSummary } from '../ProductSummary'
 import { useWindowSize } from '../../hooks/useWindowSize'
 import { MoreSidebarLayout } from '../MoreSidebarLayout'
 import { ProductProperties } from '../ProductProperties'
@@ -8,23 +9,30 @@ import { ProductMetaFields } from '../ProductMetaFields'
 import { Personalization } from '../Personalization'
 import { ProductExtras } from '../ProductExtras'
 import { ProductGallery } from '../ProductGallery'
+import { ProductMainDetails } from '../ProductMainDetails'
 
 import {
   Container
 } from './styles'
-export const ProductDetails = (props) => {
+
+const ProductDetailsUI = (props) => {
   const {
     open,
     business,
     onClose,
-    product,
+    productState,
+    formState,
     handleUpdateBusinessState,
     setFormTaxState,
     formTaxState,
     taxes,
     setTaxes,
     fees,
-    setFees
+    setFees,
+    handlechangeImage,
+    handleChangeInput,
+    handleUpdateClick,
+    handleChangeFormState
   } = props
   const { width } = useWindowSize()
   const [isMenuOpen, setIsMenuOpen] = useState(false)
@@ -89,12 +97,12 @@ export const ProductDetails = (props) => {
   useEffect(() => {
     setExtraOpen(false)
     setIsExtendExtraOpen(false)
-  }, [product])
+  }, [props.product])
 
   return (
     <Container id='product_details'>
       {(!isExtendExtraOpen || width < 1000) && (
-        <ProductMainDetails
+        <ProductSummary
           {...props}
           actionSidebar={actionSidebar}
           showOption={showOption}
@@ -106,10 +114,20 @@ export const ProductDetails = (props) => {
           isExtendExtraOpen={isExtendExtraOpen}
           onClose={handleCloseExtraOpen}
         >
+          {showOption === 'product_details' && (
+            <ProductMainDetails
+              product={productState?.product}
+              formState={formState}
+              handlechangeImage={handlechangeImage}
+              handleChangeFormState={handleChangeFormState}
+              handleChangeInput={handleChangeInput}
+              handleUpdateClick={handleUpdateClick}
+            />
+          )}
           {showOption === 'properties' && (
             <ProductProperties
               business={business}
-              product={product}
+              product={productState.product}
               handleUpdateBusinessState={handleUpdateBusinessState}
               setFormTaxState={setFormTaxState}
               formTaxState={formTaxState}
@@ -122,14 +140,14 @@ export const ProductDetails = (props) => {
           {showOption === 'ingredients' && (
             <ProductIngredient
               business={business}
-              product={product}
+              product={productState.product}
               handleUpdateBusinessState={handleUpdateBusinessState}
             />
           )}
           {showOption === 'product_options' && (
             <ProductExtras
               business={business}
-              product={product}
+              product={productState.product}
               setIsExtendExtraOpen={setIsExtendExtraOpen}
               handleUpdateBusinessState={handleUpdateBusinessState}
             />
@@ -137,16 +155,16 @@ export const ProductDetails = (props) => {
           {showOption === 'product_images' && (
             <ProductGallery
               business={business}
-              categoryId={product.category_id}
-              product={product}
+              categoryId={productState.product.category_id}
+              product={productState.product}
               handleUpdateBusinessState={handleUpdateBusinessState}
             />
           )}
           {showOption === 'custom_fields' && (
             <ProductMetaFields
               businessId={business.id}
-              categoryId={product.category_id}
-              productId={product.id}
+              categoryId={productState.product.category_id}
+              productId={productState.product.id}
             />
           )}
           {showOption === 'personalization' && (
@@ -158,4 +176,12 @@ export const ProductDetails = (props) => {
       )}
     </Container>
   )
+}
+
+export const ProductDetails = (props) => {
+  const productDetailsProps = {
+    ...props,
+    UIComponent: ProductDetailsUI
+  }
+  return <ProductDetatilsController {...productDetailsProps} />
 }
