+++ src/components/ProductMainDetails/index.js
@@ -1,221 +1,58 @@
 import React, { useState } from 'react'
-import { useLanguage, useUtils, ProductDetatils as ProductDetatilsController } from 'ordering-components-admin'
-import BsChevronRight from '@meronex/icons/bs/BsChevronRight'
-import { XLg, ThreeDots, Laptop, Phone } from 'react-bootstrap-icons'
-import { Switch } from '../../styles/Switch'
-import { ProductDetatilsEditForm } from '../ProductDetatilsEditForm'
-import { Button, IconButton } from '../../styles/Buttons'
-import { DropdownButton, Dropdown } from 'react-bootstrap'
-import { useTheme } from 'styled-components'
-import { Confirm } from '../Confirm'
-import { Modal } from '../Modal'
-import { ProductDesktopPreview } from '../ProductDesktopPreview'
+import { useLanguage } from 'ordering-components-admin'
+import { DragScroll } from '../DragScroll'
+import { ProductDetatilsInformation } from '../ProductDetatilsInformation'
 
 import {
-  ProductDetailsContainer,
-  DetailsHeader,
-  LeftHeader,
-  RightHeader,
-  ProductName,
-  ProductImage,
-  ProductDetailsContent,
-  ProductPrice,
-  ProductDescription,
-  ProductConfigsContainer,
-  ProductConfigOption,
-  ActionSelectorWrapper,
-  ProductPreviewHeader
+  Container,
+  TabsConatiner,
+  Tab
 } from './styles'
 
-const ProductMainDetailsUI = (props) => {
+export const ProductMainDetails = (props) => {
   const {
-    actionSidebar,
-    showOption,
-    handleShowOption,
-    handleChangeProductActiveState,
-    productState,
-    productCart,
+    product,
     formState,
     handlechangeImage,
     handleChangeInput,
-    handleUpdateClick,
-    handleDeleteProduct,
-    showProductOption
+    handleChangeFormState,
+    handleUpdateClick
   } = props
-
   const [, t] = useLanguage()
-  const theme = useTheme()
-  const [{ optimizeImage, parsePrice }] = useUtils()
-  const [isEditMode, setIsEditMode] = useState(false)
-  const [confirm, setConfirm] = useState({ open: false, content: null, handleOnAccept: null })
-  const [isProductPreview, setIsProductPreview] = useState(false)
-  const [selectedView, setSelectedView] = useState('desktop')
 
-  const configsOptions = [
-    {
-      key: 'properties',
-      value: t('PROPERTIES', 'Properties')
-    },
-    {
-      key: 'ingredients',
-      value: t('INGREDIENTS', 'Ingredients')
-    },
-    {
-      key: 'product_options',
-      value: t('PRODUCT_OPTIONS', 'Product options')
-    },
-    {
-      key: 'product_images',
-      value: t('PRODUCT_IMAGES', 'Product images')
-    },
-    {
-      key: 'custom_fields',
-      value: t('CUSTOM_FIELDS', 'Custom fields')
-    },
-    {
-      key: 'personalization',
-      value: t('PERSONALIZATION', 'Personalization')
-    }
+  const [selectedOption, setSelectedOption] = useState('information')
+  const listOptions = [
+    { key: 'information', content: t('INFORMATION', 'Information') },
+    { key: 'advanced', content: t('ADVANCED', 'Advanced') },
+    { key: 'labels', content: t('LABELS', 'Labels') },
+    { key: 'seo', content: t('SEO_OPTIONS', 'SEO options') }
   ]
-
-  const handleDeleteClick = () => {
-    setConfirm({
-      open: true,
-      content: t('QUESTION_DELETE_PRODUCT', 'Are you sure that you want to delete this product?'),
-      handleOnAccept: () => {
-        handleDeleteProduct()
-        setConfirm({ ...confirm, open: false })
-      }
-    })
-  }
   return (
-    <>
-      <ProductDetailsContainer>
-        <DetailsHeader>
-          <LeftHeader>
-            <ProductName>
-              {productState?.product?.name}
-            </ProductName>
-            <Switch
-              defaultChecked={productState?.product?.enabled || false}
-              onChange={handleChangeProductActiveState}
-            />
-          </LeftHeader>
-          <RightHeader>
-            <ActionSelectorWrapper>
-              <DropdownButton
-                className='product_actions'
-                menuAlign={theme?.rtl ? 'left' : 'right'}
-                title={<ThreeDots />}
-                id={theme?.rtl ? 'dropdown-menu-align-left' : 'dropdown-menu-align-right'}
-              >
-                <Dropdown.Item
-                  onClick={() => setIsProductPreview(true)}
-                >
-                  {t('PREVIEW', 'Preview')}
-                </Dropdown.Item>
-                <Dropdown.Item
-                  onClick={() => handleDeleteClick()}
-                >
-                  {t('DELETE', 'Delete')}
-                </Dropdown.Item>
-              </DropdownButton>
-            </ActionSelectorWrapper>
-            <IconButton
-              color='black'
-              onClick={() => isEditMode ? setIsEditMode(false) : actionSidebar(false)}
+    <Container>
+      <h1>{t('PRODUCT_DETAILS', 'Product details')}</h1>
+      <TabsConatiner>
+        <DragScroll>
+          {listOptions.map(option => (
+            <Tab
+              key={option.key}
+              active={selectedOption === option.key}
+              onClick={() => setSelectedOption(option.key)}
             >
-              <XLg />
-            </IconButton>
-          </RightHeader>
-        </DetailsHeader>
-        {!isEditMode ? (
-          <>
-            <ProductImage bgimage={optimizeImage(productState?.product?.images, 'h_200,c_limit')} />
-            <ProductDetailsContent>
-              <ProductPrice>{parsePrice(productState?.product?.price)}</ProductPrice>
-              <ProductDescription>{productState?.product?.description}</ProductDescription>
-              <ProductConfigsContainer>
-                {configsOptions.map(config => (
-                  <ProductConfigOption
-                    key={config.key}
-                    active={showOption === config.key}
-                    onClick={() => handleShowOption(config.key)}
-                  >
-                    <span>{config.value}</span>
-                    <BsChevronRight />
-                  </ProductConfigOption>
-                ))}
-              </ProductConfigsContainer>
-            </ProductDetailsContent>
-            <Button
-              color='secundaryDark'
-              borderRadius='7.6px'
-              onClick={() => setIsEditMode(true)}
-            >
-              {t('EDIT', 'Edit')}
-            </Button>
-          </>
-        ) : (
-          <ProductDetatilsEditForm
-            onCancel={() => setIsEditMode(false)}
-            product={productState?.product}
-            formState={formState}
-            handlechangeImage={handlechangeImage}
-            handleChangeInput={handleChangeInput}
-            handleButtonUpdateClick={handleUpdateClick}
-          />
-        )}
-      </ProductDetailsContainer>
-      <Confirm
-        title={t('WEB_APPNAME', 'Ordering')}
-        width='700px'
-        content={confirm.content}
-        acceptText={t('ACCEPT', 'Accept')}
-        open={confirm.open}
-        onClose={() => setConfirm({ ...confirm, open: false })}
-        onCancel={() => setConfirm({ ...confirm, open: false })}
-        onAccept={confirm.handleOnAccept}
-        closeOnBackdrop={false}
-      />
-
-      <Modal
-        width='900px'
-        open={isProductPreview}
-        onClose={() => setIsProductPreview(false)}
-      >
-        <ProductPreviewHeader>
-          <h1>{t('PREVIEW', 'Preview')}</h1>
-          <div>
-            <IconButton
-              color={selectedView === 'desktop' ? 'primary' : 'black'}
-              onClick={() => setSelectedView('desktop')}
-            >
-              <Laptop />
-            </IconButton>
-            <IconButton
-              color={selectedView === 'mobile' ? 'primary' : 'black'}
-              onClick={() => setSelectedView('mobile')}
-            >
-              <Phone />
-            </IconButton>
-          </div>
-        </ProductPreviewHeader>
-        <ProductDesktopPreview
-          isMobileView={selectedView === 'mobile'}
-          product={productState?.product}
-          productCart={productCart}
-          showProductOption={showProductOption}
+              {option.content}
+            </Tab>
+          ))}
+        </DragScroll>
+      </TabsConatiner>
+      {selectedOption === 'information' && (
+        <ProductDetatilsInformation
+          product={product}
+          formState={formState}
+          handlechangeImage={handlechangeImage}
+          handleChangeInput={handleChangeInput}
+          handleChangeFormState={handleChangeFormState}
+          handleButtonUpdateClick={handleUpdateClick}
         />
-      </Modal>
-    </>
+      )}
+    </Container>
   )
-}
-
-export const ProductMainDetails = (props) => {
-  const productDetailsProps = {
-    ...props,
-    UIComponent: ProductMainDetailsUI
-  }
-  return <ProductDetatilsController {...productDetailsProps} />
 }
